+++ src/lib/metadata.ts
@@ -5,4 +5,12 @@ export const siteConfig = {
     it: 'it_IT',
     en: 'en_US',
   },
+  author: {
+    jobTitle: 'Senior Software Developer',
+    email: 'me@eneascaccabarozzi.xyz',
+  },
+  social: {
+    github: 'https://github.com/enea-scaccabarozzi',
+    linkedin: 'https://www.linkedin.com/in/enea-scaccabarozzi-9660ba266',
+  },
 } as const